requirements.txt
Django==1.9.1
MySQL-python==1.2.5
argparse==1.2.1
decorator==4.0.6
dj-database-url==0.3.0
djangorestframework==3.3.2
gunicorn==19.4.5
ipython==4.0.3
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5
requests==2.9.1
simplegeneric==0.8.1
traitlets==4.1.0
wsgiref==0.1.2
